+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - core/.env : `$ENCRYPTION` environment variable can be defined to 1 to enable encryption
 - import_recordings : added verbose option to enable `INFO` log level (default is set to WARNING)
 - launch_wrapper.sh: `$EXPECTED_NODE` is now defined by an env variable.
+- camera_node : added `no_compression` to disable jpeg compression when saving images
+- bag_converter.py : updated script to export non-compressed images
 
 ### Modified